safari-534.53-branch/LayoutTests/http/tests/security/mixedContent/redirect-http-to-https-script-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that loads an insecure script (but with a
tricky redirect).  We should trigger a mixed content callback because an active
network attacker could have redirected the script load to
https://attacker.com.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-redirect-http-to-https-script.html";
></iframe>
</body>
</html>
